Major refactor of the css structure
This change install a real, updatable, maintainable for our shared less code.
This brings parity with the js structure, where core and lib folders are not project related.

Also, in order to help with performance, lib and core now comes bundled into their respective bundle.less file, in order to reduce the number of ajax request less.js needs to make when in dev.