Fix the broken bual build
The build was working in the browser but failed with grunt.

The problem arose when developers started to use classes from the
bundle.css file as less mixin, which is not possible when building the
css from only less file.

The fix is to use less import specifier to tell it to not process the
bundle.css file as less but only output it.

We also took the liberty to set core, dev and build as reference only,
since they should not output anything.