Invert the usebuild event logic (#166)
Instead of opting-in the builds, use them by default.
Let developers opt-in and make it simplier for everyone else.

Always set the use-build value in the xml for yes unless ?use-dev is prevent in the url or the use-dev cookie is present.
I also added an opt-out option with ?use-dev=no. It will set the cookie on the client for time() - 3600
Picked from 80ac5b44b2

Use only the host domain without subdomain for the cookie and restricted the button rendering for developers only
Picked from fabb4605be

Replace use-build with use-dev:
Also, make the cookie valid for all .288dev.com subdomains