Do not reject numbers
Since numbers are ignored and never added to the dictionary, we need to
make sure no rule is removed because one of its component is a number.

We do not want to had numbers into the dictionary to prevent it from
getting too large. SVG contains a bunch a number dans this would slow
all dictionary look ups.