Refactor of everything related to sharing metas.
Concept: Every string used to output the meta tags should ultimatly be a
global variable is site-variable so it can be overwriten easily in a
page.  No need to overwrite complicated templates just to change a
simple string.

og and twitter values fallback on meta-value witch compute the right
values from the page-value or site-value. If you want custom metas for a
any use-case, just overwrite the page-value.
Is also ready for specific facebook or other social media values.